backend/pyproject.toml
[project]
name = "news-analysis-backend"
version = "0.1.0"
description = "FastAPI backend for News Analysis MVP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web/API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",

    # Parsing and HTTP (reused from code/sentimen_analiz + parser)
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "newspaper3k>=0.2.8",

    # ML stack — опираемся на те же версии, что и в code/sentimen_analiz
    "transformers[torch]>=4.30.0",
    "torch==2.4.0+cpu",
    "torchvision==0.19.0+cpu",
    "torchaudio==2.4.0+cpu",
    "tokenizers>=0.13.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "nltk>=3.8.1",
    "protobuf>=3.20.0",
    "safetensors>=0.3.0",
    "huggingface-hub>=0.16.0",
    "accelerate>=1.11.0",
    "sentencepiece>=0.2.1",
    "tqdm>=4.66.5",

    # Pydantic + env
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
python-preference = "system"

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }
torchaudio = { index = "pytorch-cpu" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
